Some Changes to the script to make it work better
I think I solved the problem that a lot of people were having, first of all, I forgot to mention that you have to install the pillow package in the terminal with pip install pillow.  But, the reason why it is not clicking the skip button is that the program has a hard time detecting the images on the screen because they are very small. My solution to this is using something called "confidence", its a parameter in the locateCenterOnScreen function that instead of finding the exact image on the screen it will find an image that resembles it. But it also tends to have false positives where it will detect another image that is not the skip ad button and it'll try to press it. The solution to this is having pyautogui look through a certain region of the screen to find the skip ad button, this makes it not able to detect anything outside of the region but only the region where the skip ad button is. Also, upon detecting the skip ad button, it might click it twice because it runs the function again and detects the skip ad button again which also pauses the video, which isn't good. So I added a little delay after the clicking the skip ad button so it doesn't click twice.